+++ db/models/template.go
@@ -1,37 +1,56 @@
 package models
 
 import (
-	"errors"
-	"path"
-	"strconv"
+	"time"
 
-	"gitlab.com/codebox4073715/codebox/config"
+	dbconn "gitlab.com/codebox4073715/codebox/db/connection"
 	"gorm.io/gorm"
 )
 
 type WorkspaceTemplate struct {
-	gorm.Model
-	ID          uint   `gorm:"primarykey"`
-	Name        string `gorm:"size:255;unique;not null;"`
-	Type        string `gorm:"size:255;"`
-	Description string
-	Icon        string `gorm:"type:text;"`
+	ID          uint           `gorm:"primarykey"  json:"id"`
+	Name        string         `gorm:"size:255;unique;not null;"  json:"name"`
+	Type        string         `gorm:"size:255;"  json:"type"`
+	Description string         `json:"description"`
+	Icon        string         `gorm:"type:text;" json:"icon"`
+	CreatedAt   time.Time      `gorm:"index" json:"-"`
+	UpdatedAt   time.Time      `gorm:"index" json:"-"`
+	DeletedAt   gorm.DeletedAt `gorm:"index" json:"-"`
 }
 
-type WorkspaceTemplateVersion struct {
-	gorm.Model
-	ID             uint `gorm:"primarykey"`
-	TemplateID     uint
-	Template       *WorkspaceTemplate `gorm:"constraint:OnDelete:CASCADE;not null;"`
-	Name           string             `gorm:"size:255;not null;"`
-	ConfigFilePath string             `gorm:"type:text;" json:"config_file_relative_path"`
-	Files          string             `gorm:"type:text;not null;"`
+// Retrieve workspace template by id
+// return nil if object is not found
+func RetrieveWorkspaceTemplateByID(id uint) (*WorkspaceTemplate, error) {
+	var wt *WorkspaceTemplate
+	err := dbconn.DB.Model(WorkspaceTemplate{}).Find(&wt, map[string]interface{}{
+		"id": id,
+	}).Error
+	return wt, err
 }
 
-func (wtv *WorkspaceTemplateVersion) GetConfigFileAbsPath() (p string, err error) {
-	if wtv.ID <= 0 || wtv.Files == "" {
-		return "", errors.New("object does not exist")
+// Retrieve workspace template by name
+// return nil if object is not found
+func RetrieveWorkspaceTemplateByName(name string) (*WorkspaceTemplate, error) {
+	var wt *WorkspaceTemplate
+	err := dbconn.DB.Model(WorkspaceTemplate{}).Find(&wt, map[string]interface{}{
+		"name": name,
+	}).Error
+	return wt, err
+}
+
+// create new template
+func CreateWorkspaceTemplate(name string, templateType string, description string, icon string) (*WorkspaceTemplate, error) {
+	wt := WorkspaceTemplate{
+		Name:        name,
+		Type:        templateType,
+		Description: description,
+		Icon:        icon,
 	}
-	p = path.Join(config.Environment.UploadsPath, "git-sources", strconv.Itoa(int(wtv.ID)))
-	return p, nil
+
+	err := dbconn.DB.Save(&wt).Error
+	if err != nil {
+		return nil, err
+	}
+
+	return &wt, nil
 }